FNF specific: Add option to force loop delay of idle anims to 0.
Works by looking for the word "idle" in the animation name.
All of these would be detected by this variable:
* Dad idle dance
* dadidledance
etc.

For now (and probably never) it won't need any other fancy detection or ways to handle edge cases as it's impossible to handle everything, especially when it's for a very specific use case like here, the app isn't meant only for FNF sprites so I don't want to clutter it with 999 different conditions.