Global/Spritesheet/Animation settings are now handled by SettingsManager
This change reduces the ridiculous amount of variables needed to be passed through the various classes and functions by storing the different setting methods in dictionaries.

We're also renaming several functions to a more direct and obvious naming scheme.

Meanwhile we're also cleaning up error messages which will also get an overhaul soon.