Get and update global settings on various actions.
Instead of only updating global settings when processing starts we now update it whenever the create_override_settings_window is called.

This allows us to pre-fill animation or spritesheet override settings with the global values if none were found in animation_settings or spritesheet_settings dictionaries.
As a bonus this also allows the preview gif function to be shown with the global settings if they've been changed instead of hardcoded defaults.